requirements.txt
discord.py[voice]
asyncio
numpy
yt-dlp
